While using FVerticalWheelPicker, I succeeded in entering scores with decimal points at one point in IeltsRecordScreen.

+++ app/src/main/java/com/example/proenglishscoretracker/data/EnglishInfoRepository.kt
@@ -167,11 +167,11 @@ class EnglishInfoRepository(private val dao: EnglishInfoDao) {
     }
     suspend fun saveIeltsInfo(
         date: String,
-        readingScore: Int,
-        listeningScore: Int,
-        writingScore: Int,
-        speakingScore: Int,
-        overallScore: Int,
+        readingScore: Float,
+        listeningScore: Float,
+        writingScore: Float,
+        speakingScore: Float,
+        overallScore: Float,
         memo: String
     ) {
         val ieltsInfo = EnglishTestInfo.IELTS(
